Refactors InfiniteScrollBox component, refactors useLoadOnScrollToBottom hook
