+++ src/hooks/InfiniteScroller.ts
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { ReactNode, useEffect, useRef, useState } from 'react';
 
 const LOAD_NEXT_BOUNDARY_THRESHOLD = 80;
 
@@ -12,6 +12,15 @@ export interface UseLoadOnScrollToBottomOptions {
   }) => void;
   shouldLoadOnScroll?: boolean;
   invertScrollDirection?: boolean;
+  paging?: boolean;
+  dataElements?: ReactNode[];
+  dataElementLength?: number;
+  focusedElementIndex?: number;
+  keyboardFocusElement?: HTMLElement | HTMLElement[] | null;
+  onChangeFocusedDataElement?: (dataElementIndex: number) => void;
+  onSelectDataElement?: (dataElementIndex: number) => void;
+  onClose?: () => void;
+  enableKeyboardNavigationWrapping?: boolean;
 }
 
 /**
@@ -24,14 +33,161 @@ export const useLoadOnScrollToBottom = ({
   onChangeScrollLength,
   shouldLoadOnScroll = true,
   invertScrollDirection = false,
+  dataElementLength,
+  paging = false,
+  dataElements,
+  onSelectDataElement,
+  onClose,
+  focusedElementIndex: focusedElementIndexProp,
+  onChangeFocusedDataElement,
+  enableKeyboardNavigationWrapping = false,
+  keyboardFocusElement: keyboardFocusElementProp,
 }: UseLoadOnScrollToBottomOptions) => {
   // Refs
   const loadRef = useRef(load);
   const onChangeScrollLengthRef = useRef(onChangeScrollLength);
+  const isInitialMountRef = useRef(true);
+  const onChangeFocusedDataElementRef = useRef(onChangeFocusedDataElement);
+  const onSelectDataElementRef = useRef(onSelectDataElement);
+  const onCloseRef = useRef(onClose);
   useEffect(() => {
     loadRef.current = load;
     onChangeScrollLengthRef.current = onChangeScrollLength;
-  }, [load, onChangeScrollLength]);
+    onChangeFocusedDataElementRef.current = onChangeFocusedDataElement;
+    onSelectDataElementRef.current = onSelectDataElement;
+    onCloseRef.current = onClose;
+  }, [
+    load,
+    onChangeFocusedDataElement,
+    onChangeScrollLength,
+    onClose,
+    onSelectDataElement,
+  ]);
+
+  const limit = (() => {
+    return (
+      (() => {
+        if (element?.offsetHeight && dataElementLength) {
+          return Math.ceil(element.offsetHeight / dataElementLength);
+        }
+        return 0;
+      })() + 1
+    );
+  })();
+
+  const [offset, setOffset] = useState(0);
+
+  const [focusedElementIndex, setFocusedElementIndex] = useState(
+    focusedElementIndexProp
+  );
+
+  useEffect(() => {
+    if (focusedElementIndexProp != null) {
+      setFocusedElementIndex(focusedElementIndexProp);
+    }
+  }, [focusedElementIndexProp]);
+
+  useEffect(() => {
+    if (!isInitialMountRef.current && focusedElementIndex != null) {
+      onChangeFocusedDataElementRef.current &&
+        onChangeFocusedDataElementRef.current(focusedElementIndex);
+    }
+  }, [focusedElementIndex]);
+
+  useEffect(() => {
+    const keyboardFocusElements: HTMLElement[] = [];
+    if (keyboardFocusElementProp) {
+      if (Array.isArray(keyboardFocusElementProp)) {
+        keyboardFocusElements.push(...keyboardFocusElementProp);
+      } else {
+        keyboardFocusElements.push(keyboardFocusElementProp);
+      }
+    }
+    if (element) {
+      keyboardFocusElements.push(element);
+    }
+    if (dataElements && dataElementLength && keyboardFocusElements.length > 0) {
+      const keydownCallback = (event: KeyboardEvent) => {
+        setFocusedElementIndex((prevFocusedElementIndex = -1) => {
+          const nextFocusedOptionIndex = (() => {
+            switch (event.key) {
+              case 'ArrowUp':
+                event.preventDefault();
+                if (
+                  enableKeyboardNavigationWrapping ||
+                  !loadRef.current ||
+                  prevFocusedElementIndex > 0
+                ) {
+                  return (
+                    (!!prevFocusedElementIndex
+                      ? prevFocusedElementIndex
+                      : dataElements.length) - 1
+                  );
+                }
+                return 0;
+              case 'ArrowDown':
+                event.preventDefault();
+                return (prevFocusedElementIndex + 1) % dataElements.length;
+              case 'Enter':
+              case 'Tab':
+                if (
+                  prevFocusedElementIndex != null &&
+                  onSelectDataElementRef.current
+                ) {
+                  event.preventDefault();
+                  onSelectDataElementRef.current(prevFocusedElementIndex);
+                }
+                break;
+              case 'Escape':
+                if (onCloseRef.current) {
+                  event.preventDefault();
+                  onCloseRef.current();
+                }
+                break;
+            }
+          })();
+          if (nextFocusedOptionIndex != null && element?.offsetHeight) {
+            if (element) {
+              const nextFocusedOptionScrollTop =
+                nextFocusedOptionIndex * dataElementLength;
+              const { scrollTop: scrollUpperBound } = element;
+              const scrollLowerBound = scrollUpperBound + element.offsetHeight;
+
+              if (
+                nextFocusedOptionScrollTop + dataElementLength >
+                scrollLowerBound
+              ) {
+                element.scrollTop =
+                  scrollUpperBound +
+                  nextFocusedOptionScrollTop +
+                  dataElementLength -
+                  scrollLowerBound;
+              }
+              if (nextFocusedOptionScrollTop < scrollUpperBound) {
+                element.scrollTop = nextFocusedOptionScrollTop;
+              }
+              return nextFocusedOptionIndex;
+            }
+          }
+          return prevFocusedElementIndex >= 0 ? prevFocusedElementIndex : 0;
+        });
+      };
+      keyboardFocusElements.forEach((element) => {
+        element.addEventListener('keydown', keydownCallback);
+      });
+      return () => {
+        keyboardFocusElements.forEach((element) => {
+          element.removeEventListener('keydown', keydownCallback);
+        });
+      };
+    }
+  }, [
+    dataElementLength,
+    dataElements,
+    element,
+    enableKeyboardNavigationWrapping,
+    keyboardFocusElementProp,
+  ]);
 
   useEffect(() => {
     if (element && shouldLoadOnScroll) {
@@ -58,6 +214,9 @@ export const useLoadOnScrollToBottom = ({
             scrollTop,
             scrollLeft,
           });
+        if (dataElementLength) {
+          setOffset(Math.floor(scrollTop / dataElementLength));
+        }
       };
       element.addEventListener('scroll', scrollEventCallback);
       scrollEventCallback();
@@ -65,5 +224,42 @@ export const useLoadOnScrollToBottom = ({
         element.removeEventListener('scroll', scrollEventCallback);
       };
     }
-  }, [bottomThreshold, element, invertScrollDirection, shouldLoadOnScroll]);
+  }, [
+    bottomThreshold,
+    dataElementLength,
+    element,
+    invertScrollDirection,
+    shouldLoadOnScroll,
+  ]);
+
+  useEffect(() => {
+    isInitialMountRef.current = false;
+    return () => {
+      isInitialMountRef.current = true;
+    };
+  }, []);
+
+  const displayableDataSet = (() => {
+    if (dataElements) {
+      return paging ? dataElements.slice(offset, offset + limit) : dataElements;
+    }
+  })();
+
+  const dataElementsContainerHeight = (() => {
+    if (paging && dataElementLength && dataElements) {
+      return dataElements.length * dataElementLength;
+    }
+  })();
+
+  const dataElementsContainerFillerElementHeight = (() => {
+    if (dataElementLength) {
+      return offset * dataElementLength;
+    }
+  })();
+
+  return {
+    dataElementsContainerHeight,
+    dataElementsContainerFillerElementHeight,
+    displayableDataSet,
+  };
 };
